Move animations to separate folder (#83)
