+++ src/core/animations/seqcombinator.hpp
@@ -1,5 +1,5 @@
-#ifndef CORE_SEQCOMBINATOR_HPP_
-#define CORE_SEQCOMBINATOR_HPP_
+#ifndef CORE_ANIMATIONS_SEQCOMBINATOR_HPP_
+#define CORE_ANIMATIONS_SEQCOMBINATOR_HPP_
 
 #include <vector>
 #include <initializer_list>
@@ -70,4 +70,4 @@ private:
     State m_emptyPlaceholder;
 };
 
-#endif  // CORE_SEQCOMBINATOR_HPP_
+#endif  // CORE_ANIMATIONS_SEQCOMBINATOR_HPP_
